requirements.txt
Django==5.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
google-auth==2.27.0
python-decouple==3.8
psycopg2-binary==2.9.9
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
dj-database-url==2.1.0
textstat>=0.7.3
python-Levenshtein>=0.25.0
xhtml2pdf>=0.2.16
